pyproject.toml
[project]
name = "gitlab-scripts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.117.1",
    "langchain-community>=0.3.29",
    "langchain-openai>=0.3.33",
    "pandas>=2.3.2",
    "python-gitlab>=6.3.0",
    "uvicorn>=0.37.0",
]
